requirements.txt
botogram2
wit
python-dotenv
itranslate
python-dateutil